Convert C to native C++ code.
Make current_mailbox and current_mailbox_acl members of imapscaninfo.
Add a constructor that takes a pointer to an existing imapscaninfo
object and constructs a new one with the same maiblox and ACL. This
is not a copy constructor, since nothing else gets copied.

Remove the explicit maildir parameter from imapscan_maildir() and
imapscan_openfile(), they'll read it from the imapscaninfo object.

Replace C strings with std::strings, in several places. A NULL
char * becomes an empty string. This is also used for error checking,
but the errors are fatal errors, and they can be handled as a validation
failure (denied ACL, in case of strings that contain ACL rights).

Move the check for an open folder into doNoop, instead of checking for
it before calling doNoop.

Do not open an inotify file descriptor immediately upon constructing
a watch object, instead create it, on demand, on the first open.